+++ WordPress/ruleset.xml
@@ -91,4 +91,23 @@
 		<severity>0</severity>
 	</rule>
 
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.XSS.EscapeOutput.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.XSS.EscapeOutput.OutputNotEscaped">
+		<severity>0</severity>
+	</rule>
+	<rule ref="WordPress.XSS.EscapeOutput.OutputNotEscapedShortEcho">
+		<severity>0</severity>
+	</rule>
+	<rule ref="WordPress.XSS.EscapeOutput.UnsafePrintingFunction">
+		<severity>0</severity>
+	</rule>
+	<rule ref="WordPress.XSS.EscapeOutput.DeprecatedCustomSanitizingFunctions">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
